# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: OGDEN Atlas API
  description: |
    Geospatial land intelligence API for the OGDEN Land Design Atlas.
    Manages projects, data layers, design features, PDF exports, public portals,
    real-time collaboration (WebSocket), and AI enrichment.
  version: 0.2.0
  contact:
    name: OGDEN Atlas
    url: https://atlas.ogden.ag
servers:
  - url: /api/v1
    description: API v1
security:
  - bearerAuth: []
...
paths:
...
  /layers/project/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Layers]
      summary: List all data layers for a project (any project role)
      responses:
        '200':
          description: Layer list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
  /layers/project/{projectId}/{layerType}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: layerType
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Layers]
      summary: Get a specific data layer (any project role)
      responses:
        '200':
          description: Layer data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '404':
          description: Layer not found
  /layers/project/{projectId}/{layerType}/refresh:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: layerType
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Layers]
      summary: Re-fetch layer from source API (any project role)
      responses:
        '202':
          description: Refresh job queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          jobId:
                            type: string
                            format: uuid

  # ── Design Features ────────────────────────────────────────────────────────
  /design-features/project/{projectId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Design Features]
      summary: List all design features for a project (any project role)
      responses:
        '200':
          description: Feature list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/DesignFeature'

    post:
      tags: [Design Features]
      summary: Create a design feature (owner or designer)
      description: Broadcasts a feature_created WebSocket event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesignFeatureInput'
      responses:
        '201':
          description: Feature created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DesignFeature'
  /design-features/project/{projectId}/{featureType}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: featureType
        in: path
        required: true
        schema:
          type: string
          enum: [zone, structure, path, point, annotation]
    get:
      tags: [Design Features]
      summary: List features filtered by type (any project role)
      responses:
        '200':
          description: Filtered feature list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
  /design-features/project/{projectId}/bulk:
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Design Features]
      summary: Bulk insert design features (owner or designer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [features]
              properties:
                features:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateDesignFeatureInput'
      responses:
        '201':
          description: Features inserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
  /design-features/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      tags: [Design Features]
      summary: Update a design feature (owner or designer)
      description: Broadcasts a feature_updated WebSocket event.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subtype:
                  type: string
                label:
                  type: string
                properties:
                  type: object
                phaseTag:
                  type: string
                geometry:
                  type: object
                style:
                  type: object
                sortOrder:
                  type: integer
      responses:
        '200':
          description: Feature updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '404':
          description: Feature not found

    delete:
      tags: [Design Features]
      summary: Delete a design feature (owner only)
      description: Broadcasts a feature_deleted WebSocket event.
      responses:
        '204':
          description: Feature deleted
        '404':
          description: Feature not found

  # ── Exports ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/register or /auth/login (expires in 7 days)
...
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta:
          type: object
          properties:
            total:
              type: integer
        error:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
...
    DesignFeature:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        featureType:
          type: string
          enum: [zone, structure, path, point, annotation]
        subtype:
          type: string
          nullable: true
        geometry:
          type: object
          description: GeoJSON geometry
        label:
          type: string
          nullable: true
        properties:
          type: object
        phaseTag:
          type: string
          nullable: true
        style:
          type: object
          nullable: true
        sortOrder:
          type: integer
        createdBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateDesignFeatureInput:
      type: object
      required: [featureType, geometry, properties, sortOrder]
      properties:
        featureType:
          type: string
          enum: [zone, structure, path, point, annotation]
        subtype:
          type: string
        geometry:
          type: object
        label:
          type: string
        properties:
          type: object
        phaseTag:
          type: string
        style:
          type: object
        sortOrder:
          type: integer
...